more modular, before integrating more animations

+++ Assets/Scripts/Gameplay/Health.cs
@@ -2,8 +2,8 @@ using UnityEngine;
 
 public class Health : MonoBehaviour
 {
-    public int maxHealth = 100;
-    private int currentHealth;
+    public float maxHealth = 100;
+    private float currentHealth;
 
     // on death event
     public System.Action onDeath;
@@ -13,7 +13,7 @@ public class Health : MonoBehaviour
         currentHealth = maxHealth;
     }
 
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
         Debug.Log($"{gameObject.name} took {amount} damage.");
         currentHealth -= amount;
